+++ react/__tests__/PaginationDots.test.tsx
@@ -1,7 +1,9 @@
 import React from 'react'
 import { render, fireEvent } from '@vtex/test-tools/react'
 
-import PaginationDots from '../components/PaginationDots'
+import PaginationDots, { CSS_HANDLES } from '../components/PaginationDots'
+import { useContextCssHandles } from '../modules/cssHandles'
+import { mockedUseContextCssHandlesFn } from '../__fixtures__/CssHandlesHelper'
 
 const mockGoForward = jest.fn()
 const mockGoBack = jest.fn()
@@ -40,6 +42,18 @@ jest.mock('../hooks/useSliderControls', () => {
     }),
   }
 })
+
+jest.mock('../modules/cssHandles', () => ({
+  useContextCssHandles: jest.fn(),
+}))
+
+const mockedUseContextCssHandles = useContextCssHandles as jest.Mock<
+  ReturnType<typeof useContextCssHandles>
+>
+
+mockedUseContextCssHandles.mockImplementation(() =>
+  mockedUseContextCssHandlesFn(CSS_HANDLES)
+)
 
 describe('Basic rendering', () => {
   it('should render the correct number of pagination dots based on the number of slider pages', () => {