+++ react/__tests__/Slider.test.tsx
@@ -1,8 +1,10 @@
 import React from 'react'
 import { render } from '@vtex/test-tools/react'
 
-import Slider from '../components/Slider'
+import Slider, { CSS_HANDLES } from '../components/Slider'
 import { mockInitialInfiniteSliderState } from '../__fixtures__/SliderStateContext'
+import { mockedUseContextCssHandlesFn } from '../__fixtures__/CssHandlesHelper'
+import { useContextCssHandles } from '../modules/cssHandles'
 
 let mockDeviceDetectorReturn = { device: 'desktop', isMobile: false }
 
@@ -22,6 +24,18 @@ jest.mock('../components/SliderContext', () => ({
 jest.mock('vtex.device-detector', () => ({
   useDevice: jest.fn(() => mockDeviceDetectorReturn),
 }))
+
+jest.mock('../modules/cssHandles', () => ({
+  useContextCssHandles: jest.fn(),
+}))
+
+const mockedUseContextCssHandles = useContextCssHandles as jest.Mock<
+  ReturnType<typeof useContextCssHandles>
+>
+
+mockedUseContextCssHandles.mockImplementation(() =>
+  mockedUseContextCssHandlesFn(CSS_HANDLES)
+)
 
 describe('Basic rendering', () => {
   it('should render complete slider, with arrows and pagination dots on default settings', () => {