+++ react/components/Slider.tsx
@@ -1,16 +1,25 @@
 import React, { FC, useRef, Fragment, ReactNode } from 'react'
 import { useDevice } from 'vtex.device-detector'
-import { useCssHandles } from 'vtex.css-handles'
 
 import { useScreenResize } from '../hooks/useScreenResize'
 import { useTouchHandlers } from '../hooks/useTouchHandlers'
 import { useAutoplay } from '../hooks/useAutoplay'
-import { useSliderState } from './SliderContext'
-import SliderTrack from './SliderTrack'
-import Arrow from './Arrow'
-import PaginationDots from './PaginationDots'
+import {
+  SliderLayoutProps,
+  SliderLayoutSiteEditorProps,
+  useSliderState,
+} from './SliderContext'
+import SliderTrack, {
+  CSS_HANDLES as SliderTrackCssHandles,
+} from './SliderTrack'
+import Arrow, { CSS_HANDLES as ArrowCssHandles } from './Arrow'
+import PaginationDots, {
+  CSS_HANDLES as PaginationDotsCssHandles,
+} from './PaginationDots'
+import { useContextCssHandles } from '../modules/cssHandles'
 
 interface Props extends SliderLayoutSiteEditorProps {
+  arrowSize: number
   totalItems: number
   itemsPerPage: number
   centerMode: SliderLayoutProps['centerMode']
@@ -19,7 +28,13 @@ interface Props extends SliderLayoutSiteEditorProps {
   children?: Array<Exclude<ReactNode, boolean | null | undefined>>
 }
 
-const CSS_HANDLES = ['sliderLayoutContainer', 'sliderTrackContainer'] as const
+export const CSS_HANDLES = [
+  'sliderLayoutContainer',
+  'sliderTrackContainer',
+  ...SliderTrackCssHandles,
+  ...ArrowCssHandles,
+  ...PaginationDotsCssHandles,
+] as const
 
 const Slider: FC<Props> = ({
   children,
@@ -33,7 +48,7 @@ const Slider: FC<Props> = ({
   itemsPerPage,
   centerMode,
 }) => {
-  const handles = useCssHandles(CSS_HANDLES)
+  const { handles } = useContextCssHandles()
   const { isMobile } = useDevice()
   const { label = 'slider', slidesPerPage } = useSliderState()
   const containerRef = useRef<HTMLDivElement>(null)