+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Prop `classes` to SliderLayout.
+
+### Changed
+- Migrate to `vtex.css-handles@1.x`.
 
 ## [0.16.0] - 2020-12-16
 ### Added